Move template processing code to its own package
The pkg/template package is now the parent to more than just
template processing. Templates are their own resource and there are
other related resources such as template instances. It makes sense for
template processing code to move to its own package.

This commit does the following:
- updates the doc.go in pkg/template
- moves the Processor type and and related tests to
  pkg/template/processor
- moves the registry storage class for template processing to its own
  package: pkg/template/registry/processor
- leaves non-processing parameter-related utility functions along with
  corresponding tests in pkg/template
