+++ pkg/template/apiserver/apiserver.go
@@ -13,7 +13,7 @@ import (
 
 	templateapiv1 "github.com/openshift/origin/pkg/template/apis/template/v1"
 	brokertemplateinstanceetcd "github.com/openshift/origin/pkg/template/registry/brokertemplateinstance/etcd"
-	templateregistry "github.com/openshift/origin/pkg/template/registry/template"
+	templateprocessor "github.com/openshift/origin/pkg/template/registry/processor"
 	templateetcd "github.com/openshift/origin/pkg/template/registry/template/etcd"
 	templateinstanceetcd "github.com/openshift/origin/pkg/template/registry/templateinstance/etcd"
 )
@@ -104,7 +104,7 @@ func (c *TemplateConfig) newV1RESTStorage() (map[string]rest.Storage, error) {
 	}
 
 	v1Storage := map[string]rest.Storage{}
-	v1Storage["processedTemplates"] = templateregistry.NewREST()
+	v1Storage["processedTemplates"] = templateprocessor.NewREST()
 	v1Storage["templates"] = templateStorage
 	v1Storage["templateinstances"] = templateInstanceStorage
 	v1Storage["templateinstances/status"] = templateInstanceStatusStorage
