+++ pkg/template/doc.go
@@ -1,3 +1,6 @@
-// Package template provides TemplateProcessor, capable of
-// transforming Template objects into Config objects.
+// Package template is the parent of packages related
+// to templates, including the template resource registry, template
+// instances, and template processing.
+// It includes generic utilities to help with parsing
+// template references, and working with templates
 package template
