+++ pkg/template/params.go
@@ -1,0 +1,26 @@
+package template
+
+import (
+	templateapi "github.com/openshift/origin/pkg/template/apis/template"
+)
+
+// AddParameter adds new custom parameter to the Template. It overrides
+// the existing parameter, if already defined.
+func AddParameter(t *templateapi.Template, param templateapi.Parameter) {
+	if existing := GetParameterByName(t, param.Name); existing != nil {
+		*existing = param
+	} else {
+		t.Parameters = append(t.Parameters, param)
+	}
+}
+
+// GetParameterByName searches for a Parameter in the Template
+// based on its name.
+func GetParameterByName(t *templateapi.Template, name string) *templateapi.Parameter {
+	for i, param := range t.Parameters {
+		if param.Name == name {
+			return &(t.Parameters[i])
+		}
+	}
+	return nil
+}
