+++ pkg/template/processor/doc.go
@@ -1,0 +1,3 @@
+// Package processor contains the Processor type which converts
+// templates + parameters into a resolved resource list.
+package processor
