+++ pkg/template/processor/template.go
@@ -1,4 +1,4 @@
-package template
+package processor
 
 import (
 	"fmt"
@@ -111,27 +111,6 @@ func stripNamespace(obj runtime.Object) {
 			}
 		}
 	}
-}
-
-// AddParameter adds new custom parameter to the Template. It overrides
-// the existing parameter, if already defined.
-func AddParameter(t *templateapi.Template, param templateapi.Parameter) {
-	if existing := GetParameterByName(t, param.Name); existing != nil {
-		*existing = param
-	} else {
-		t.Parameters = append(t.Parameters, param)
-	}
-}
-
-// GetParameterByName searches for a Parameter in the Template
-// based on its name.
-func GetParameterByName(t *templateapi.Template, name string) *templateapi.Parameter {
-	for i, param := range t.Parameters {
-		if param.Name == name {
-			return &(t.Parameters[i])
-		}
-	}
-	return nil
 }
 
 // EvaluateParameterSubstitution replaces escaped parameters in a string with values from the
