+++ pkg/template/registry/processor/rest_test.go
@@ -1,4 +1,4 @@
-package template
+package processor
 
 import (
 	"testing"
